shape mechanism done

+++ src/cs3500/animator/controller/TextFieldListener.java
@@ -25,7 +25,7 @@ public class TextFieldListener implements ActionListener {
 
     if (text.equals("Change the speed of the animation by entering a new speed value here.")) {
       String s = jtx.getText();
-      double speed = Double.parseDouble(s);
+      int speed = Integer.parseInt(s);
       controller.speedChanged(speed);
     }
     else {
